pyproject.toml
[build-system]
build-backend="setuptools.build_meta"
requires=["setuptools == 67.1.0", "wheel"]

[project]
authors=[
    {name="The InvokeAI Project", email="lincoln.stein@gmail.com"},

    {name="Kyle Schouviller", email="kyle0654@hotmail.com"},
    {name="Matthias Wild", email="mauwii@outlook.de"},
    {name="Younesse ANDAM", email="younesse.andam@gmail.com"},
]
dependencies=["numpy", "pillow", "tqdm"]
description='This library implements the PatchMatch based inpainting algorithm.'
dynamic=["version"]
license={file="LICENSE"}
name='PyPatchMatch'
readme={content-type="text/markdown", file="README.md"}
requires-python=">=3.9,<3.11"

[project.urls]
'Source Code'='https://github.com/mauwii/PyPatchMatch'

[tool.setuptools.dynamic]
version={attr="patchmatch.__version__"}

[tool.setuptools.packages.find]
include=["patchmatch", "patchmatch.csrc"]

[tool.setuptools.package-data]
"patchmatch"=['Makefile', 'csrc/*', 'travis.sh']

[project.optional-dependencies]
"dev"=["black", "flake8", "flake8-black", "isort", "pre-commit", "pylance"]
"dist"=[
    "auditwheel; sys_platform=='linux'",
    "build",
    "delocate; sys_platform=='darwin'",
    "pip-tools",
    "pipdeptree",
    "requests",
    "twine",
]
"test"=["pytest", "pytest-coverage"]

[tool.black]
exclude='''
/(
    .git
    | .venv
    | build
    | csrc
    | dist
)/
'''
include='.pyi?$'
line-length=88
source=['examples', 'patchmatch']
target-version=['py39']

[tool.coverage.report]
fail_under=75

[tool.isort]
profile="black"

[tool.pytest.ini_options]
addopts=["--cov-branch", "--cov-report=term:skip-covered", "--cov=patchmatch"]
minversion="6.0"
python_files=["test_*.py"]
testpaths=["tests"]
